pyproject.toml
[project]
name = "modality-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.4.0",
    "altair>=5.5.0",
    "arviz>=0.20.0",
    "bitsandbytes>=0.45.5",
    "datasets>=3.3.2",
    "en-core-web-sm",
    "matplotlib>=3.10.1",
    "numba>=0.61.0",
    "outlines>=0.2.1",
    "polars>=1.29.0",
    "protobuf>=6.31.1",
    "pymc>=5.21.0",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "sentencepiece>=0.2.0",
    "spacy>=3.8.7",
    "statsmodels>=0.14.4",
    "transformers>=4.49.0",
]

[project.optional-dependencies]
cuda = [
    "flash-attn",
]

[tool.uv]
no-build-isolation-package = ["flash-attn"]

[tool.uv.sources]
flash-attn = { url = "https://github.com/Zarrac/my-pytorch-builds/releases/download/flash-attn-2.7.4.post1-cuda12.8/flash_attn-2.7.4.post1+pt270cu128cxx11abiTRUE-cp312-cp312-linux_x86_64.whl" }
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }

[project.scripts]
modality-llm = "modality_llm.cli:main_cli"

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pandas-stubs>=2.2.3.250527",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "types-requests>=2.32.0.20250602",
    "types-tqdm>=4.67.0.20250516",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/modality_llm"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --doctest-modules --cov-report term-missing:skip-covered --cov=modality_llm"
testpaths = [
    "tests",
    "src",
]
doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL"

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src", "tests"]
